requirements.txt
flask==3.1.2
flask-cors==6.0.2
gunicorn==23.0.0
pdf2docx==0.5.8
pillow==12.0.0
openpyxl==3.1.5
pandas==2.3.3
PyPDF2==3.0.1
python-docx==1.2.0
stripe==14.1.0
python-dotenv==1.2.1
reportlab==4.4.6

# AI/ML Libraries
easyocr==1.7.2
pytesseract==0.3.13
opencv-python==4.10.0.84
transformers==4.47.1
torch==2.6.0
scikit-learn==1.6.0
langdetect==1.0.9
deep-translator==1.11.4
nltk==3.9.1
textblob==0.18.0.post0
numpy==1.26.4
scipy==1.14.1

# For model caching and optimization
huggingface-hub==0.27.0
sentencepiece==0.2.0
